PolAux

Documentation

type PolF2 = Vect F2 (Lex String)

The data type PolF2 is the field of polynomial with coefficients in the finite field F2.

expTo1 :: PolF2 -> PolF2

(expTo1 p) is the representative with leastest degree of the polynomial p in the quotient group F2[x_1,...,x_N]/(x_1+x_1^2,...,x_N+x_N^2). The main idea is to replace every ocurrence of x_i^M with x_i thus we obtain an identical polynomial without exponents greater than 1.

In the library HaskellForMaths exists a function that performs the same (%%) so we can check the results. For example,

>>> expTo1 (x1^3)
x1
>>> (x1^3) %% [x1^2+x1]
x1
>>> expTo1 (x1^3*x2^6+x3^2*x4+x1+1)
x1x2+x1+x3x4+1
>>> let pol = x1^3*x2^6+x3^2*x4+x1+1
>>> let list = [x1^2+x1,x2^2+x2,x3^2+x3,x4^2+x4]
>>> pol %% list
x1x2+x1+x3x4+1

deriv :: PolF2 -> PolF2 -> PolF2

(deriv p v) is the derivative of the polynomial p with respect to the variable v. It's important to note that deriv only works if it applies to polynomials without exponents greater than 1. In practice, deriv will only be used with the polynomials that have been previously embeded in the quotient group described above. For example, >>> let v = x1 >>> deriv v v 1 >>> deriv (1+x1+x2+x1*x2) v x2+1 >>> deriv (x1*x2+x1+x3*x4+1) v x2+1

var :: (Num k, MonomialConstructor m) => v -> Vect k (m v)

vars :: (Num k, Ord k, MonomialConstructor m, Ord (m v)) => Vect k (m v) -> [Vect k (m v)]

zerov :: Vect k b

mindices :: MonomialConstructor m => forall v. m v -> [(v, Int)]

lm :: Vect b c -> c

lt :: Vect t1 t -> (t, t1)