ReadingF

Documentation

clause2pol :: [String] -> (PolF2, Set PolF2)

(clause2pol cs) is a pair (p,vs), where p is the polynomial that corresponds to the clause cs (which is written in DIMACS format) and vs is the set of its variables.

>>> clause2pol ["1"]
(x1,fromList [x1])
>>> clause2pol ["1","-2"]
(x1x2+x2+1,fromList [x1,x2])

var' :: String -> (PolF2, PolF2)

(var' lit) is the pair (p,v), where p is the polynomial and v is the variable that corresponds (if it does) to the literal lit (in DIMACS format).

>>> var' "0"
(0,0)
>>> var' "1"
(x1,x1)
>>> var' "-1"
(x1+1,x1)

variable2List :: (t, Set a) -> (t, [a])

(variable2List (a,b)) returns a pair which first element remains unchanged and the second is the list of elements of set b. For example,

>>> variable2List (1,S.empty)
(1,[])
>>> variable2List (x1,S.fromList[0,2])
(x1,[0,2])

insertPol :: (PolF2, Set PolF2) -> (Set PolF2, Set PolF2) -> (Set PolF2, Set PolF2)

(insertPol (a,b) (acc,vs)) is a pair which first element is set acc plus element a and the second is the union of set vs and set b. For example,

>>> insertPol (x1,S.fromList[x1]) (S.empty,S.empty)
(fromList [x1],fromList [x1])
>>> insertPol (x1*x2,S.fromList[x1,x2]) (S.fromList[x1],S.fromList[x1])
(fromList [x1x2,x1],fromList [x1,x2])
>>> insertPol (x1,S.fromList[x1]) (S.fromList[x1*x2],S.fromList[x1,x2])
(fromList [x1x2,x1],fromList [x1,x2])

hasV :: PolF2 -> PolF2 -> Int

(hasV v p) is 1 if p has the variable v otherwise it is 0. For example,

>>> hasV x1 x1
1
>>> hasV x2 (x1+1)
0
>>> hasV x1 (x1*x2+1)
1

counting :: (Set PolF2, [PolF2]) -> [(PolF2, Int)] -> [(PolF2, Int)]

(counting (ps,vvs) cs) is a list of ordered pairs in which the first element is a variable from vvs and the second indicates in how many polynomials of the set ps this variable occurs. For example,

>>> counting (S.empty, []) []
[]
>>> counting (S.fromList [x1*x2,x1,x1+1,x2], [x1,x2]) []
[(x2,2),(x1,3)]
>>> counting (S.fromList [x1], [x1,x2,x3]) []
[(x3,0),(x2,0),(x1,1)]

dimacs2pols :: FilePath -> IO ()

(dimacs2pols f) is the pair (ps,vs) where ps is the set of polynomials wich corresponds to the formula in DIMACS format writed in the file f and vs is the list of variables wich occurs in any polynomial.

>>> dimacs2pols "exDIMACS/easy/example1.txt"
(fromList [x1x2+x1+x2,1],[x2,x1])
>>> dimacs2pols "exDIMACS/easy/example2.txt"
(fromList [x1x2+x1+x2,x1x2+x1+1,1],[x2,x1])
>>> dimacs2pols "exDIMACS/easy/example3.txt"
(fromList [x1x2+x1+x2,x1x2+x1+1,x1x2+x2+1,1],[x2,x1])
>>> dimacs2pols "exDIMACS/easy/example4.txt"
(fromList [x1x2+x1+x2,x1x2+x1+1,x1x2+x2+1,x1x2+1,1],[x2,x1])

dimacs2pols' :: FilePath -> IO ()

dimacs2polsSize :: FilePath -> IO ()

(dimacs2pols f) is the pair (n,m) where n is the size of the set of polynomials wich corresponds to the formula in DIMACS format writed in the file f and m is the length of the list of variables wich occurs in any polynomial.

>>> dimacs2polsSize "exDIMACS/easy/example1.txt"
(2,2)
>>> dimacs2polsSize "exDIMACS/easy/example2.txt"
(3,2)
>>> dimacs2polsSize "exDIMACS/easy/example3.txt"
(4,2)
>>> dimacs2polsSize "exDIMACS/easy/example4.txt"
(5,2)
>>> dimacs2polsSize "exDIMACS/hard/sat100.cnf"
(431,100)
>>> dimacs2polsSize "exDIMACS/hard/sat250.cnf"
(1066,250)

main :: FilePath -> IO ()

(main f) is verified if the set of formulas in DIMACS format in the file f were satisfiable. Otherwise, (main f) would return False.

>>> main "exDIMACS/easy/example1.txt"
True
>>> main "exDIMACS/easy/example2.txt"
True
>>> main "exDIMACS/easy/example3.txt"
True
>>> main "exDIMACS/easy/example4.txt"
False
>>> main "exDIMACS/medium/exampleSat0.txt"
True
>>> main "exDIMACS/medium/exampleSat1.txt"
True

mainCount :: FilePath -> IO ()

(mainCount f) returns a pair where first component is verified if the set of formulas in DIMACS format in the file f were satisfiable; and the second component counts how many times the delta rule has been executed.

>>> mainCount "exDIMACS/easy/example1.txt"
(True,1)
>>> mainCount "exDIMACS/easy/example2.txt"
(True,3)
>>> mainCount "exDIMACS/easy/example3.txt"
(True,7)
>>> mainCount "exDIMACS/easy/example4.txt"
(False,12)
>>> mainCount "exDIMACS/medium/exampleSat0.txt"
(True,114)
>>> mainCount "exDIMACS/medium/exampleSat1.txt"
(True,142)

mainTrace :: [Expr] -> [PolF2] -> IO ()