Tool

Description

This module aims to provide an efficient tool to solve the SAT problem. Both the algorithm and the implementation have been carried out by the researchers from the department of Computer Science and Artificial Intelligence at the University of Seville.

Documentation

tool :: (Eq u, Show (m u), Show u, MonomialConstructor m, Algebra k (m u), Ord (m u), Ord k, Num k) => [Vect k (m u)] -> Bool

tool decides if the set of formulas that produced the set of polynomials were satisfiables. The function input is a list of polynomials because the transformation from formula to polynomial is handled by ReadingF.hs module.

varsList :: (Foldable t, Num k, MonomialConstructor m, Ord (m v), Ord k) => t (Vect k (m v)) -> [Vect k (m v)]

varsList xs return the set of variables which occurs in any polynomial of the list xs. For example, the trivial case:

>>> let xs = (take 10 (map (var . (\n -> 'x':n) . show) [1..])) :: [LexPoly F2 String]
>>> xs
[x1,x2,x3,x4,x5,x6,x7,x8,x9,x10]
>>> varsList xs
[x1,x2,x3,x4,x5,x6,x7,x8,x9,x10]
>>> varsList [x1*x2+x2,x1+x1*x3,x1*x4*x2]
[x1,x2,x3,x4]