ToolCount

Description

This module aims to provide an efficient tool to solve the SAT problem. Both the algorithm and the implementation have been carried out by the researchers from the department of Computer Science and Artificial Intelligence at the University of Seville.

Documentation

toolCount :: Integer -> (Set PolF2, Set PolF2) -> (Bool, Integer)

(tool (ps,vvs)) is verified if the original set of formulas which polynomials from ps came was satisfiable. Otherwise, the function will return False if that set of formulas was unsatisfiable. Note that vvs is the set of variables which occurs in any polynomial from ps. For example,

toolLC :: Integer -> (Set PolF2, [PolF2]) -> (Bool, Integer)