ToolL

Description

This module aims to provide an efficient tool to solve the SAT problem. Both the algorithm and the implementation have been carried out by the researchers from the department of Computer Science and Artificial Intelligence at the University of Seville.

Documentation

toolL :: (Set PolF2, [PolF2]) -> Bool

(tool (ps,vvs)) is verified if the original set of formulas which polynomials from ps came was satisfiable. Otherwise, the function will return False if that set of formulas was unsatisfiable. Note that vvs is the set of variables which occurs in any polynomial from ps. For example,

>>> tool (S.fromList[1],S.empty)
True
>>> tool (S.fromList[x1,x1+1],S.fromList[x1])
False

deltaRule :: PolF2 -> PolF2 -> PolF2 -> PolF2

(deltaRule v p1 p2) is the polynomial obtained from apply the independence rule described in the paper [?] between polynomials p1 p2 respect the variable v. It's important to note that as v is the variable from wich we derive, it won't occurs in the output polynomial. For example,

>>> deltaRule x1 1 1
1
>>> deltaRule x1 1 0
0
>>> deltaRule x1 x1 x1
1

deltaRule1Step :: PolF2 -> Set PolF2 -> Set PolF2 -> Set PolF2

(deltaRule1Step v pps acum) is the union of the set accum, which ramains invariant, and the set of polynomials obtained after applying deltaRule respect v between every polynomial from set pps. For example,

>>> deltaRule1Step x1 (S.fromList [x1]) (S.fromList [1])
fromList [1]
>>> deltaRule1Step x1 (S.fromList [x1,x1*x2,x1*x3]) (S.empty)
fromList [x2x3,x2,x3,1]

nextL :: PolF2 -> Set PolF2 -> Set PolF2

nextLRec :: (Num t, Eq t) => t -> (Set PolF2, [Vect F2 (Lex String)]) -> (Set PolF2, [Vect F2 (Lex String)])