ToolS

Description

This module aims to provide an efficient tool to solve the SAT problem. Both the algorithm and the implementation have been carried out by the researchers from the department of Computer Science and Artificial Intelligence at the University of Seville.

Documentation

varsSet :: (Ord (m v), Num k, Ord k, MonomialConstructor m) => Vect k (m v) -> Set (Vect k (m v))

varsSet returns the set of variables which occurs in a polynomial. For example: λ> varsSet (x2*x3+x1) fromList [x1,x2,x3] λ> varsSet (x1) fromList [x1]

varsList :: (Foldable t, Num k, MonomialConstructor m, Ord (m v), Ord k) => t (Vect k (m v)) -> Set (Vect k (m v))

varsList xs returns the set of variables which occurs in any polynomial of the set xs. For example:

deltaRule :: (Eq k, Eq u, Num k, Ord (m u), Algebra k (m u), MonomialConstructor m, Show (m u), Show u) => Vect k (m u) -> Vect k (m u) -> Vect k (m u) -> Vect k (m u)

The function (deltaRule p x y) performs the independence rule described in the paper [?]. It's important to note that p is the variable from wich we derive and the one we would drop. For example: >>> deltaRule (x1:: LexPoly F2 String) (1:: LexPoly F2 String) (1:: LexPoly F2 String) 1 >>> deltaRule (x1:: LexPoly F2 String) (1:: LexPoly F2 String) (0:: LexPoly F2 String) 0 >>> deltaRule (x1:: LexPoly F2 String) (x1:: LexPoly F2 String) (x1:: LexPoly F2 String) 1

deltaRule1Step :: (Eq k, Eq u, Num k, Ord k, Ord (m u), Algebra k (m u), MonomialConstructor m, Show (m u), Show u) => Vect k (m u) -> Set (Vect k (m u)) -> Set (Vect k (m u)) -> Set (Vect k (m u))

deltaRule1Step apply deltaRule from p between every polynomial in the first list and store the results in the accumulator (second list). For exaple: λ> deltaRuleList1Step (x1:: LexPoly F2 String) ([x1]:: [LexPoly F2 String]) ([1]::[LexPoly F2 String]) [1,1] λ> deltaRuleList1Step (x1:: LexPoly F2 String) ([x1,x1*x2,x1*x3]:: [LexPoly F2 String]) ([]::[LexPoly F2 String]) [x3,x2x3,x2,x3,x2,1]