PolAux

Documentation

clean :: (Eq k, Num k, Eq (m u), Show u, Ord (m u), Show (m u), Algebra k (m u), MonomialConstructor m) => Vect k (m u) -> Vect k (m u)

clean aims to select the simplest representative of a polynomial in the quotient group F2[x_1,...,x_N]/(x_1+x_1^2,...,x_N+x_N^2). The main idea is to replace every ocurrence of x_i^M with x_i thus we obtain an identical polynomial without exponents greater than 1.

In the library HaskellForMaths exists a function that performs the same (%%) so we can check the results. For example,

>>> clean (x1^3:: LexPoly F2 String)
x1
>>> (x1^3:: LexPoly F2 String) %% [x1^2+x1]
x1
>>> clean (x1^3*x2^6+x3^2*x4+x1+1 :: LexPoly F2 String)
x1x2+x1+x3x4+1
>>> let pol = x1^3*x2^6+x3^2*x4+x1+1 :: LexPoly F2 String
>>> let list = [x1^2+x1,x2^2+x2,x3^2+x3,x4^2+x4]
>>> pol %% list
x1x2+x1+x3x4+1

deriv :: (Eq k, Eq u, Num k, Ord (m u), Show (m u), Algebra k (m u), MonomialConstructor m) => Vect k (m u) -> Vect k (m u) -> Vect k (m u)

deriv calculates the derivative of the polynomial f with respect to the variable v. It's important to note that we use the fact that we are in the quotient group described above. For example:

>>> let v = x1:: LexPoly F2 String
>>> deriv v v
1
>>> deriv (1+x1+x2+x1*x2:: LexPoly F2 String) v
x2+1
>>> deriv (x1*x2+x1+x3*x4+1:: LexPoly F2 String) v
x2+1