ReadingF

Documentation

clause2pol :: [String] -> (PolF2, Set PolF2)

(clause2pol cs) is a pair (p,vs), where p is the polynomial that corresponds to the clause cs (which is written in DIMACS format) and vs is the set of its variables.

>>> clause2pol ["1"]
(x1,fromList [x1])
>>> clause2pol ["1","-2"]
(x1x2+x2+1,fromList [x1,x2])

var' :: String -> (PolF2, PolF2)

(var' lit) is the pair (p,v), where p is the polynomial and v is the variable that corresponds (if it does) to the literal lit (in DIMACS format).

>>> var' "0"
(0,0)
>>> var' "1"
(x1,x1)
>>> var' "-1"
(x1+1,x1)

dimacs2pols :: FilePath -> IO ()

(dimacs2pols f) is the pair (ps,vs) where ps is the set of polynomials wich corresponds to the formula in DIMACS format writed in the file f and vs is the set of variables wich occurs in any polynomial.

>>> dimacs2pols "exDIMACS/easy/example1.txt"
(fromList [x1x2+x1+x2,1],fromList [x1,x2])
>>> dimacs2pols "exDIMACS/easy/example4.txt"
(fromList [x1x2+x1+x2,x1x2+x1+1,x1x2+x2+1,x1x2+1,1],fromList [x1,x2])

main1 :: FilePath -> IO ()

(main f) is verified if the set of formulas in DIMACS format in the file f were satisfiable. Otherwise, (main f) would return False.

>>> main1 "exDIMACS/easy/example1.txt"
True
>>> main1 "exDIMACS/easy/example4.txt"
False

insertPol :: (Ord a, Ord a1) => (a1, Set a) -> (Set a1, Set a) -> (Set a1, Set a)

variable2List :: (t, Set a) -> (t, [a])

mainList :: FilePath -> IO ()

var'' :: String -> PolF2

clause2pol' :: [String] -> PolF2

mainN :: FilePath -> IO ()

leer :: FilePath -> IO ()

fileSat0 :: [[Char]]

fileSat00' :: Set PolF2

fileSat00 :: (Set PolF2, Set PolF2)

fileSat1 :: [[Char]]